debian/rules: Use G_DISABLE_ASSERT to disable libnice asserts
If not disabled, libnice will abort if at any moment the state machine
goes to an unexpected state, or any other error happens.

This is not desired in a server where there might be hundreths or
thousands of connections going on, and throwing down the whole server
would be a VERY bad thing to do just because one of those connections
happens to go into a bad state.